ask the aggregators alone at step 2
Plus an optional country before the walk; the chip editor is a static module both pages mount.
